refactor, adopt in notebook and hover contribution
