+++ src/vs/editor/contrib/hover/browser/hoverContribution.ts
@@ -12,12 +12,8 @@ import { MarkdownHoverParticipant } from 'vs/editor/contrib/hover/browser/markdo
 import { MarkerHoverParticipant } from 'vs/editor/contrib/hover/browser/markerHoverParticipant';
 import { HoverController } from 'vs/editor/contrib/hover/browser/hoverController';
 import 'vs/css!./hover';
-import { AccessibleViewImplentation, AccessibleViewRegistry } from 'vs/platform/accessibility/browser/accessibleViewRegistry';
-import { EditorContextKeys } from 'vs/editor/common/editorContextKeys';
-import { ICodeEditorService } from 'vs/editor/browser/services/codeEditorService';
-import { AccessibleViewProviderId, AccessibleViewType, IAccessibleViewService } from 'vs/platform/accessibility/browser/accessibleView';
-import { IContextViewService } from 'vs/platform/contextview/browser/contextView';
-import { IHoverService } from 'vs/platform/hover/browser/hover';
+import { AccessibleViewRegistry } from 'vs/platform/accessibility/browser/accessibleViewRegistry';
+import { ExtHoverAccessibleView, HoverAccessibleView } from 'vs/editor/contrib/hover/browser/hoverAccessibleViews';
 
 registerEditorContribution(HoverController.ID, HoverController, EditorContributionInstantiation.BeforeFirstInteraction);
 registerEditorAction(ShowOrFocusHoverAction);
@@ -44,51 +40,5 @@ registerThemingParticipant((theme, collector) => {
 		collector.addRule(`.monaco-editor .monaco-hover hr { border-bottom: 0px solid ${hoverBorder.transparent(0.5)}; }`);
 	}
 });
-const implementation: AccessibleViewImplentation = {
-	priority: 95, name: 'hover', implementation: accessor => {
-		const accessibleViewService = accessor.get(IAccessibleViewService);
-		const codeEditorService = accessor.get(ICodeEditorService);
-		const editor = codeEditorService.getActiveCodeEditor() || codeEditorService.getFocusedCodeEditor();
-		const editorHoverContent = editor ? HoverController.get(editor)?.getWidgetContent() ?? undefined : undefined;
-		if (!editor || !editorHoverContent) {
-			return false;
-		}
-		const language = editor?.getModel()?.getLanguageId() ?? 'typescript';
-		accessibleViewService.show({
-			id: AccessibleViewProviderId.Hover,
-			verbositySettingKey: 'accessibility.verbosity.hover',
-			provideContent() { return editorHoverContent; },
-			onClose() {
-				HoverController.get(editor)?.focus();
-			},
-			options: { language, type: AccessibleViewType.View }
-		});
-		return true;
-	}, when: EditorContextKeys.hoverFocused
-};
-const extImplementation: AccessibleViewImplentation = {
-	priority: 90, name: 'extension-hover', implementation: accessor => {
-		const accessibleViewService = accessor.get(IAccessibleViewService);
-		const contextViewService = accessor.get(IContextViewService);
-		const contextViewElement = contextViewService.getContextViewElement();
-		const extensionHoverContent = contextViewElement?.textContent ?? undefined;
-		const hoverService = accessor.get(IHoverService);
-
-		if (contextViewElement.classList.contains('accessible-view-container') || !extensionHoverContent) {
-			// The accessible view, itself, uses the context view service to display the text. We don't want to read that.
-			return false;
-		}
-		accessibleViewService.show({
-			id: AccessibleViewProviderId.Hover,
-			verbositySettingKey: 'accessibility.verbosity.hover',
-			provideContent() { return extensionHoverContent; },
-			onClose() {
-				hoverService.showAndFocusLastHover();
-			},
-			options: { language: 'typescript', type: AccessibleViewType.View }
-		});
-		return true;
-	}
-};
-AccessibleViewRegistry.registerImplementation(implementation);
-AccessibleViewRegistry.registerImplementation(extImplementation);
+AccessibleViewRegistry.registerImplementation(new HoverAccessibleView());
+AccessibleViewRegistry.registerImplementation(new ExtHoverAccessibleView());
