+++ src/vs/workbench/browser/workbench.ts
@@ -48,6 +48,8 @@ import { setHoverDelegateFactory } from 'vs/base/browser/ui/hover/hoverDelegateF
 import { setBaseLayerHoverDelegate } from 'vs/base/browser/ui/hover/hoverDelegate2';
 import { AccessibilityProgressSignalScheduler } from 'vs/platform/accessibilitySignal/browser/progressAccessibilitySignalScheduler';
 import { setProgressAcccessibilitySignalScheduler } from 'vs/base/browser/ui/progressbar/progressAccessibilitySignal';
+import { AccessibleViewRegistry } from 'vs/platform/accessibility/browser/accessibleViewRegistry';
+import { NotificationAccessibleView } from 'vs/workbench/browser/parts/notifications/notificationAccessibleView';
 
 export interface IWorkbenchOptions {
 
@@ -415,6 +417,9 @@ export class Workbench extends Layout {
 
 		// Register Commands
 		registerNotificationCommands(notificationsCenter, notificationsToasts, notificationService.model);
+
+		// Register notification accessible view
+		AccessibleViewRegistry.registerImplementation(new NotificationAccessibleView());
 
 		// Register with Layout
 		this.registerNotifications({
