+++ src/vs/workbench/contrib/accessibility/browser/accessibleViewContributions.ts
@@ -3,31 +3,15 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
-/*---------------------------------------------------------------------------------------------
- *  Copyright (c) Microsoft Corporation. All rights reserved.
- *  Licensed under the MIT License. See License.txt in the project root for license information.
- *--------------------------------------------------------------------------------------------*/
-
 import { Disposable } from 'vs/base/common/lifecycle';
 import { ICodeEditorService } from 'vs/editor/browser/services/codeEditorService';
-import { localize } from 'vs/nls';
 import { IKeybindingService } from 'vs/platform/keybinding/common/keybinding';
 import { AccessibilityVerbositySettingId, accessibleViewIsShown } from 'vs/workbench/contrib/accessibility/browser/accessibilityConfiguration';
 import * as strings from 'vs/base/common/strings';
-import { ICommandService } from 'vs/platform/commands/common/commands';
-import { getNotificationFromContext } from 'vs/workbench/browser/parts/notifications/notificationsCommands';
-import { IListService, WorkbenchList } from 'vs/platform/list/browser/listService';
-import { NotificationFocusedContext } from 'vs/workbench/common/contextkeys';
-import { alert } from 'vs/base/browser/ui/aria/aria';
 import { AccessibilityHelpAction, AccessibleViewAction } from 'vs/workbench/contrib/accessibility/browser/accessibleViewActions';
-import { IAction } from 'vs/base/common/actions';
-import { INotificationViewItem } from 'vs/workbench/common/notifications';
-import { ThemeIcon } from 'vs/base/common/themables';
-import { Codicon } from 'vs/base/common/codicons';
 import { InlineCompletionsController } from 'vs/editor/contrib/inlineCompletions/browser/inlineCompletionsController';
 import { InlineCompletionContextKeys } from 'vs/editor/contrib/inlineCompletions/browser/inlineCompletionContextKeys';
 import { ContextKeyExpr, IContextKeyService } from 'vs/platform/contextkey/common/contextkey';
-import { AccessibilitySignal, IAccessibilitySignalService } from 'vs/platform/accessibilitySignal/browser/accessibilitySignalService';
 import { COMMENTS_VIEW_ID, CommentsMenus } from 'vs/workbench/contrib/comments/browser/commentsTreeViewer';
 import { IViewsService } from 'vs/workbench/services/views/common/viewsService';
 import { CommentsPanel, CONTEXT_KEY_HAS_COMMENTS } from 'vs/workbench/contrib/comments/browser/commentsView';
@@ -52,139 +36,29 @@ export class AccesibleViewHelpContribution extends Disposable {
 			accessor.get(IAccessibleViewService).showAccessibleViewHelp();
 			return true;
 		}, accessibleViewIsShown));
-		// TODO: @meganrogge move this out, have separate types so help or view can be contributed
 		AccessibleViewRegistry.getImplementations().forEach(impl => {
-			this._register(AccessibleViewAction.addImplementation(impl.priority, impl.name, impl.implementation, impl.when));
+			if (impl.type === AccessibleViewType.View) {
+				this._register(AccessibleViewAction.addImplementation(impl.priority, impl.name, impl.implementation, impl.when));
+			} else {
+				this._register(AccessibilityHelpAction.addImplementation(impl.priority, impl.name, impl.implementation, impl.when));
+			}
 		});
 	}
 }
 
-export class NotificationAccessibleViewContribution extends Disposable {
-	static ID: 'notificationAccessibleViewContribution';
+
+export class AccesibleViewContributions extends Disposable {
+	static ID: 'accesibleViewContributions';
 	constructor() {
 		super();
-		this._register(AccessibleViewAction.addImplementation(90, 'notifications', accessor => {
-			const accessibleViewService = accessor.get(IAccessibleViewService);
-			const listService = accessor.get(IListService);
-			const commandService = accessor.get(ICommandService);
-			const accessibilitySignalService = accessor.get(IAccessibilitySignalService);
-
-			function renderAccessibleView(): boolean {
-				const notification = getNotificationFromContext(listService);
-				if (!notification) {
-					return false;
-				}
-				commandService.executeCommand('notifications.showList');
-				let notificationIndex: number | undefined;
-				let length: number | undefined;
-				const list = listService.lastFocusedList;
-				if (list instanceof WorkbenchList) {
-					notificationIndex = list.indexOf(notification);
-					length = list.length;
-				}
-				if (notificationIndex === undefined) {
-					return false;
-				}
-
-				function focusList(): void {
-					commandService.executeCommand('notifications.showList');
-					if (list && notificationIndex !== undefined) {
-						list.domFocus();
-						try {
-							list.setFocus([notificationIndex]);
-						} catch { }
-					}
-				}
-				const message = notification.message.original.toString();
-				if (!message) {
-					return false;
-				}
-				notification.onDidClose(() => accessibleViewService.next());
-				accessibleViewService.show({
-					id: AccessibleViewProviderId.Notification,
-					provideContent: () => {
-						return notification.source ? localize('notification.accessibleViewSrc', '{0} Source: {1}', message, notification.source) : localize('notification.accessibleView', '{0}', message);
-					},
-					onClose(): void {
-						focusList();
-					},
-					next(): void {
-						if (!list) {
-							return;
-						}
-						focusList();
-						list.focusNext();
-						alertFocusChange(notificationIndex, length, 'next');
-						renderAccessibleView();
-					},
-					previous(): void {
-						if (!list) {
-							return;
-						}
-						focusList();
-						list.focusPrevious();
-						alertFocusChange(notificationIndex, length, 'previous');
-						renderAccessibleView();
-					},
-					verbositySettingKey: AccessibilityVerbositySettingId.Notification,
-					options: { type: AccessibleViewType.View },
-					actions: getActionsFromNotification(notification, accessibilitySignalService)
-				});
-				return true;
+		AccessibleViewRegistry.getImplementations().forEach(impl => {
+			if (impl.type === AccessibleViewType.View) {
+				this._register(AccessibleViewAction.addImplementation(impl.priority, impl.name, impl.implementation, impl.when));
+			} else {
+				this._register(AccessibilityHelpAction.addImplementation(impl.priority, impl.name, impl.implementation, impl.when));
 			}
-			return renderAccessibleView();
-		}, NotificationFocusedContext));
-	}
-}
-
-function getActionsFromNotification(notification: INotificationViewItem, accessibilitySignalService: IAccessibilitySignalService): IAction[] | undefined {
-	let actions = undefined;
-	if (notification.actions) {
-		actions = [];
-		if (notification.actions.primary) {
-			actions.push(...notification.actions.primary);
-		}
-		if (notification.actions.secondary) {
-			actions.push(...notification.actions.secondary);
-		}
-	}
-	if (actions) {
-		for (const action of actions) {
-			action.class = ThemeIcon.asClassName(Codicon.bell);
-			const initialAction = action.run;
-			action.run = () => {
-				initialAction();
-				notification.close();
-			};
-		}
-	}
-	const manageExtension = actions?.find(a => a.label.includes('Manage Extension'));
-	if (manageExtension) {
-		manageExtension.class = ThemeIcon.asClassName(Codicon.gear);
-	}
-	if (actions) {
-		actions.push({
-			id: 'clearNotification', label: localize('clearNotification', "Clear Notification"), tooltip: localize('clearNotification', "Clear Notification"), run: () => {
-				notification.close();
-				accessibilitySignalService.playSignal(AccessibilitySignal.clear);
-			}, enabled: true, class: ThemeIcon.asClassName(Codicon.clearAll)
 		});
 	}
-	return actions;
-}
-
-export function alertFocusChange(index: number | undefined, length: number | undefined, type: 'next' | 'previous'): void {
-	if (index === undefined || length === undefined) {
-		return;
-	}
-	const number = index + 1;
-
-	if (type === 'next' && number + 1 <= length) {
-		alert(`Focused ${number + 1} of ${length}`);
-	} else if (type === 'previous' && number - 1 > 0) {
-		alert(`Focused ${number - 1} of ${length}`);
-	}
-	return;
 }
